Conflicts:
	Makefile
	src/main/blackbox/blackbox_io.c
	src/main/config/config.c
	src/main/drivers/accgyro_mpu.c
	src/main/drivers/accgyro_mpu6500.c
	src/main/drivers/adc_stm32f30x.c
	src/main/drivers/bus_i2c_stm32f30x.c
	src/main/drivers/bus_spi.h
	src/main/drivers/compass_ak8975.c
	src/main/drivers/dma.c
	src/main/drivers/dma.h
	src/main/drivers/light_led_stm32f10x.c
	src/main/drivers/light_ws2811strip_stm32f30x.c
	src/main/drivers/nvic.h
	src/main/drivers/pwm_mapping.c
	src/main/drivers/pwm_mapping.h
	src/main/drivers/pwm_rx.c
	src/main/drivers/serial.c
	src/main/drivers/serial.h
	src/main/drivers/serial_softserial.c
	src/main/drivers/serial_uart.c
	src/main/drivers/serial_uart_impl.h
	src/main/drivers/serial_uart_stm32f10x.c
	src/main/drivers/serial_uart_stm32f30x.c
	src/main/drivers/serial_usb_vcp.c
	src/main/drivers/serial_usb_vcp.h
	src/main/drivers/sonar_hcsr04.c
	src/main/drivers/sonar_hcsr04.h
	src/main/drivers/sound_beeper_stm32f10x.c
	src/main/drivers/system.c
	src/main/drivers/timer.c
	src/main/flight/mixer.h
	src/main/flight/pid.c
	src/main/flight/pid.h
	src/main/io/beeper.c
	src/main/io/serial.c
	src/main/io/serial_cli.c
	src/main/io/serial_msp.c
	src/main/io/serial_msp.h
	src/main/main.c
	src/main/mw.c
	src/main/rx/rx.c
	src/main/rx/spektrum.h
	src/main/sensors/acceleration.c
	src/main/sensors/battery.c
	src/main/sensors/gyro.c
	src/main/sensors/gyro.h
	src/main/sensors/initialisation.c
	src/main/sensors/sonar.c
	src/main/target/CC3D/target.h
	src/main/target/NAZE/target.h
	src/main/target/SPARKY/target.h
	src/main/telemetry/hott.c
	src/main/telemetry/sport.h
	src/main/vcp/hw_config.c
	src/test/unit/battery_unittest.cc
	src/test/unit/platform.h

+++ src/main/sensors/acceleration.h
@@ -41,7 +41,7 @@ extern sensor_align_e accAlign;
 extern acc_t acc;
 extern uint16_t acc_1G;
 
-extern int16_t accADC[XYZ_AXIS_COUNT];
+extern int32_t accADC[XYZ_AXIS_COUNT];
 
 typedef struct rollAndPitchTrims_s_def {
     int16_t roll;
